+++ VivaDicta/Views/ModelsScreen/CloudModelCard.swift
@@ -2,126 +2,119 @@
 //  CloudModelCard.swift
 //  VivaDicta
 //
-//  Created by Anton Novoselov on 2025.09.03
+//  Created by Anton Novoselov on 2025.11.18
 //
 
 import SwiftUI
 
 struct CloudModelCard: View {
-    private var model: CloudModel
-    private var onConfigure: (CloudModel) -> Void
-    
+    let model: CloudModel
+    let onConfigure: (CloudModel) -> Void
+
+    @State private var selectedTab: TranscriptionModelType = .cloud
+
+    private var isRecommended: Bool {
+        // Logic to determine if model is recommended
+        model.name.contains("scribe-1") || model.name.contains("whisper-1")
+    }
+
     private var isAPIConfigured: Bool {
         model.apiKey != nil
     }
-    
-    init(model: CloudModel,
-         onConfigure: @escaping (CloudModel) -> Void) {
-        self.model = model
-        self.onConfigure = onConfigure
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(alignment: .center, spacing: 8) {
-                VStack(alignment: .leading, spacing: 8) {
-                    header
-                    metadataSection
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-                
-                actionSection
-            }
-            descriptionSection
-        }
-        .padding(16)
-        .background(.gray.opacity(0.1), in: .rect(cornerRadius: 16))
-    }
-    
-    private var header: some View {
-        HStack {
-            Text(model.displayName)
-                .font(.headline.weight(.semibold))
-            statusBadge
-            Spacer()
-        }
-    }
-    
-    private var metadataSection: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(spacing: 12) {
-                HStack(spacing: 4) {
-                    Text(model.provider.rawValue.capitalized)
-                    Image(systemName: "cloud")
-                }
-                
-                HStack(spacing: 4) {
-                    Text(model.language)
-                    Image(systemName: "globe")
-                }
-            }
-            .foregroundStyle(.secondary)
-            .font(.caption)
-            
-            HStack(spacing: 3) {
-                Text("Accuracy")
-                    .font(.caption.weight(.medium))
-                    .foregroundStyle(.secondary)
-                ModelPerformanceStatsDots(value: model.accuracy * 10)
-            }
-        }
-    }
-    
-    private var statusBadge: some View {
-        Group {
-            if !isAPIConfigured {
-                Text("Add API key")
-                    .font(.caption.weight(.medium))
-                    .padding(.horizontal, 6)
-                    .padding(.vertical, 2)
-                    .background(.orange, in: .rect(cornerRadius: 16))
-            }
-        }
-    }
-    
-    private var descriptionSection: some View {
-        Text(model.description)
-            .multilineTextAlignment(.leading)
-            .font(.caption)
-            .foregroundStyle(.secondary)
-            .lineLimit(2)
-            .fixedSize(horizontal: false, vertical: true)
-            .padding(.top, 4)
-    }
-    
-    private var actionSection: some View {
-        VStack {
-            configureButton
-        }
-        .font(.callout.weight(.semibold))
-    }
-    
-    var configureButton: some View {
-        Button {
-            onConfigure(model)
-        } label: {
-            HStack(spacing: 4) {
-                Text("Configure")
-                Image(systemName: "gear")
-            }
-        }
 
-        .foregroundStyle(.white)
-        .padding(.vertical, 4)
-        .padding(.horizontal, 6)
-        .background(.blue, in: .capsule)
+    var body: some View {
+        VStack(alignment: .leading, spacing: 20) {
+            // Header with name and badge
+            HStack(alignment: .top) {
+                VStack(alignment: .leading, spacing: 6) {
+                    Text(model.displayName)
+                        .font(.title3)
+                        .fontWeight(.semibold)
+
+                    Text(model.language)
+                        .font(.subheadline)
+                        .foregroundStyle(.secondary)
+                }
+
+                if isRecommended {
+                    Text("Recommended")
+                        .font(.caption)
+                        .fontWeight(.medium)
+                        .foregroundStyle(.blue)
+                        .padding(.horizontal, 10)
+                        .padding(.vertical, 4)
+                        .background(Color.blue.opacity(0.15))
+                        .cornerRadius(12)
+                }
+
+                Spacer()
+            }
+
+            HStack(spacing: 0) {
+                // Metrics Section
+                VStack(spacing: 8) {
+                    ModelMetricRow(
+                        label: "Speed",
+                        value: Int(model.speed * 10),
+                        color: .green
+                    )
+
+                    ModelMetricRow(
+                        label: "Accuracy",
+                        value: Int(model.accuracy * 10),
+                        color: .orange
+                    )
+                }
+
+                Spacer()
+
+                // Cloud model configuration button
+                VStack(spacing: -8) {
+                    Button(action: {
+                        onConfigure(model)
+                    }) {
+                        Image(systemName: isAPIConfigured ? "key.circle.fill" : "key.circle")
+                            .foregroundStyle(isAPIConfigured ? .green : .blue)
+                            .font(.system(size: 30))
+                            .frame(width: 56, height: 56)
+                    }
+                    Text(isAPIConfigured ? "Configured" : "Add API Key")
+                        .font(.system(size: 10))
+                        .foregroundStyle(.secondary)
+                }
+            }
+
+            // Description
+            Text(model.description)
+                .font(.subheadline)
+                .foregroundStyle(.secondary)
+                .lineLimit(4)
+                .fixedSize(horizontal: false, vertical: true)
+        }
+        .padding(20)
+        .background(.white)
+        .cornerRadius(20)
     }
 }
 
 #Preview {
-    CloudModelCard(
-        model: TranscriptionModelProvider.allCloudModels[0],
-        onConfigure: {_ in print("configure") }
-    )
-}
+    VStack(spacing: 20) {
+        // Preview with unconfigured cloud model
+        if let cloudModel = TranscriptionModelProvider.allCloudModels.first {
+            CloudModelCard(
+                model: cloudModel,
+                onConfigure: { _ in print("Configure") }
+            )
+        }
 
+        // Preview with configured cloud model (simulated)
+        if let cloudModel = TranscriptionModelProvider.allCloudModels.last {
+            CloudModelCard(
+                model: cloudModel,
+                onConfigure: { _ in print("Configure") }
+            )
+        }
+    }
+    .padding()
+    .background(.gray)
+}
